+++ src/app/pos/_components/TableBox.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@/components/common/Button";
+import Button from "@/components/common/Button/Button";
 import Icon from "@/components/common/Icon";
 import { paymentTimeTranslate } from "@/constants/translates";
 import cn from "@/lib/utils";
