+++ src/components/common/Button/Button.tsx
@@ -48,4 +48,4 @@ const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
 );
 Button.displayName = "Button";
 
-export { Button, buttonVariants };
+export default Button;
