+++ src/types/index.d.ts
@@ -21,3 +21,5 @@ declare global {
     timestamp: string;
   }
 }
+
+export {};
