+++ tsconfig.json
@@ -15,6 +15,7 @@
     "incremental": true,
     "types": ["jest"],
     "baseUrl": ".",
+    "typeRoots": ["./node_modules/@types", "./src/types"],
     "plugins": [
       {
         "name": "next"
